+++ Packages/com.mizore-nekoyanagi.utils.mizore-package-exporter/Texts/ExporterTexts.cs
@@ -159,6 +159,7 @@ namespace MizoreNekoyanagi.PublishUtil.PackageExporter {
         public static string UsePostProcessScript => Get( "UsePostProcessScript" );
         public static string UsePostProcessScriptConfirm => Get( "UsePostProcessScriptConfirm" );
         public static string PostProcessScript => Get( "PostProcessScript" );
+        public static string PostProcessScriptUpdate => Get( "PostProcessScriptUpdate" );
         public static string PostProcessScriptTooltip => Get( "PostProcessScriptTooltip" );
         public static string PostProcessScriptResetAllFields => Get( "PostProcessScriptResetAllFields" );
         public static string PostProcessScriptResetField( string fieldName ) => string.Format( Get( "PostProcessScriptResetField" ), fieldName );